game.flask/gatech/requirements.txt
Django==1.6.5
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.0.0
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2
requests==2.3.0
Flask==0.10.1
Jinja2==2.6
Werkzeug==0.8.3
Flask-SQLAlchemy==2.0
Flask-Migrate==1.3.0
Flask-Session==0.1.1
redis==2.10.3
Flask-UUID==0.2
Flask-Bootstrap==3.3.2.1
enum==0.4.4
SQLAlchemy-Utils==0.30.11
passlib==1.6.2 
